reorg(processors): remove messengers, executors, scorers
