--- ctm_ai/executors/__init__.py
@@ -1,17 +1,0 @@
-from .executor_api import APIExecutor
-from .executor_audio import AudioExecutor
-from .executor_base import BaseExecutor
-from .executor_language import LanguageExecutor
-from .executor_tool import ToolExecutor
-from .executor_video import VideoExecutor
-from .executor_vision import VisionExecutor
-
-__all__ = [
-    'BaseExecutor',
-    'AudioExecutor',
-    'LanguageExecutor',
-    'ToolExecutor',
-    'VideoExecutor',
-    'VisionExecutor',
-    'APIExecutor',
-]
