+++ ctm_ai/processors/__init__.py
@@ -1,23 +1,19 @@
-from .processor_api import APIProcessor, register_api_processors
+from .processor_api import APIProcessor
 from .processor_audio import AudioProcessor
 from .processor_base import BaseProcessor
 from .processor_code import CodeProcessor
 from .processor_language import LanguageProcessor
-from .processor_math import MathProcessor
-from .processor_tool import ToolProcessor, register_tool_processors
+from .processor_tool import ToolProcessor
 from .processor_video import VideoProcessor
 from .processor_vision import VisionProcessor
 
 __all__ = [
-    'BaseProcessor',
-    'VisionProcessor',
-    'LanguageProcessor',
-    'MathProcessor',
-    'CodeProcessor',
-    'AudioProcessor',
-    'VideoProcessor',
     'ToolProcessor',
-    'register_tool_processors',
     'APIProcessor',
-    'register_api_processors',
+    'BaseProcessor',
+    'CodeProcessor',
+    'LanguageProcessor',
+    'VisionProcessor',
+    'VideoProcessor',
+    'AudioProcessor',
 ]
