+++ ctm_ai/processors/processor_code.py
@@ -1,22 +1,20 @@
-from ..executors.executor_base import BaseExecutor
-from ..messengers.messenger_base import BaseMessenger
-from ..scorers.scorer_base import BaseScorer
+from typing import Any, Dict, List
+
 from .processor_base import BaseProcessor
 
 
 @BaseProcessor.register_processor('code_processor')
 class CodeProcessor(BaseProcessor):
-    REQUIRED_KEYS = ['OPENAI_API_KEY']
+    REQUIRED_KEYS = ['GEMINI_API_KEY']
 
-    def init_messenger(self) -> BaseMessenger:
-        return BaseMessenger.create_messenger('code_messenger')
+    def _init_info(self, *args: Any, **kwargs: Any) -> None:
+        self.system_prompt = 'You are an expert in code writing.'
 
-    def init_executor(
-        self, system_prompt: str = None, model: str = None
-    ) -> BaseExecutor:
-        return BaseExecutor(
-            name='code_executor', system_prompt=system_prompt, model=model
-        )
-
-    def init_scorer(self) -> BaseScorer:
-        return BaseScorer(name='language_scorer')
+    def build_executor_messages(
+        self,
+        query: str,
+        *args: Any,
+        **kwargs: Any,
+    ) -> List[Dict[str, Any]]:
+        self._init_info(*args, **kwargs)
+        return [{'role': 'user', 'content': query}]
