+++ ctm_ai/scorers/__init__.py
@@ -1,4 +1,3 @@
 from .scorer_base import BaseScorer
-from .scorer_language import LanguageScorer
 
-__all__ = ['BaseScorer', 'LanguageScorer']
+__all__ = ['BaseScorer']
