--- ctm_ai/scorers/scorer_language.py
@@ -1,7 +1,0 @@
-from .scorer_base import BaseScorer
-
-
-@BaseScorer.register_scorer('language_scorer')
-class LanguageScorer(BaseScorer):
-    def init_scorer(self, *args, **kwargs) -> None:
-        super().init_scorer(*args, **kwargs)
